Disable 'create group' button in home page
Disabled 'create group' button in home page for unconfirmed users.

Used partial to show alert message and disabled button.

The partial was previously in 'views/users/memberships',
used for the 'My groups' section. Now it's on 'views/application'.

Also added a method to 'integration_helpers' to test that the button
is disabled.