Fix N+1 still happening in 'My groups'
The problem was coming from the 'user_is_organizer_but_not_owner'
method in UserMembershipsHelper. Changing the way we check if the
user is the owner of the group we don't have an N+1 problem and we
don't need to include any other table in the query.

It's good to keep this in mind: Always check properly the stack trace
and find where the problem is coming from.