Implement group authorization redirections feature
With this feature, when users are logged out and visit a group
resource, they are redirected to the log in path.
When they are logged in and are not members, they are redirected to
the root path with a message explaining that they need to be members
of the group to see that resource.

The logic was extracted into the Groups::AuthorizationRedirecter
concern.

I tried to do it with a service object but I didn't like the idea of
passing the whole controller into the service object as it could
increase memory consumption and reduce performance.

The downsides of using the concern are the reduced explicitness in
the controller, and having to find the group twice for the first
action that is visited.

There has to be a better way of doing it.